\name{CORM}

\alias{CORM}

\docType{package}

\title{
  Clustering of Regression Models Method
}

\description{
  The clustering of regression models method models the expression 
  level of each gene with regression and clusters genes based on 
  the regression coefficients.
}

\details{
  \tabular{ll}{
    Package: \tab CORM\cr
    Type: \tab Package\cr
    Version: \tab 1.0.1\cr
    Date: \tab 2014-2-3\cr
    License: \tab GPL(>=2)\cr.
  }
}

\author{ Li-Xuan Qin 
  \email{qinl@mskcc.org}
}

\references{
 \itemize{
  \item Li-Xuan Qin and Steven G. Self (2006). The clustering of regression models
  method with applications in gene expression data.  \emph{Biometrics}
  62, 526-533.
  \item Li-Xuan Qin (2008). An integrative analysis of microRNA and mRNA expression - 
  a case study. \emph{Cancer Informatics} 6, 369-379.
  \item Li-Xuan Qin, Linda Breeden and Steven G. Self (2014). Finding gene cluster 
  for a replicated time course study. \emph{BMC Res Notes} 7:60.
 }
}

\keyword{Introduction}

\seealso{
  \code{\link{fit.CLM}},
  \code{\link{fit.CLMM}},
  \code{\link{fit.CLMM.2}},
  \code{\link{fit.CLMM.NA}},
  \code{\link{fit.CLMM.NA.2}}
}
